Gen_questions_hierarchial
 - new function

gen_questions_hierarchial(text)
is an amended version of gen questions.  The format as taken from the client app.js file needs to be followed in the output of components.py .

The main changes is that questions[] is now children[]  and has json style params  with unique ids, questions and answers and also children.

 - I have expanded the test case sentence to be a paragraph for further improvements

- imported uuid for unique ids (for the questions)