UPDATES FOR EASY INSTALLATION
Okay so the previous commit has some errors that would occur on a fresh install on a different machine.

I have created three installation videos - one for each module, python, server and client respectfull.

I have updated the schema for the database to include user creation as that was absent and also update the the cors cross origin allowances to include all potential origin points that the application has.